+++ Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Allow user to override default Emscripten settings in [#1059](https://github.com/PyO3/maturin/pull/1059)
+
 ## [0.13.2] - 2022-08-14
 
 * Deprecate manylinux 2010 support in [#858](https://github.com/PyO3/maturin/pull/858).